multiplicators methods

+++ src/main/java/stock/tools/interfaces/FairPriceTool.java
@@ -7,5 +7,5 @@ import java.util.List;
 public interface FairPriceTool {
 
 
-    long getFairPrice(double p1, double p2, double p3, double p4);
+    double getFairPrice(double p1, double p2, double p3, double p4);
 }
